Empty commit

+++ rassilka/api/views.py
@@ -1,7 +1,8 @@
-from rest_framework.authentication import SessionAuthentication, BasicAuthentication
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from drf_yasg.utils import swagger_auto_schema
+#from drf_yasg import openapi
 
 from .serializers import (
     CampaignSerializer, CustomerSerializer
@@ -11,6 +12,9 @@ from core.models import (
 )
 from core.stats import (
     get_stats_from_camps, get_stats_from_msgs
+)
+from .specs import (
+    campaign_par, campaign_post_request, campaign_post_responses, campaign_get_response_all, campaign_get_response, campaign_put_request, campaign_put_responses, customer_par, customer_post_request, customer_post_responses, customer_get_response_all, customer_get_response, customer_put_responses, msg_par, stats_get_response_main, stats_get_response
 )
 
 
@@ -30,7 +34,30 @@ def json_response(msg='', succeed=1):
             'error': msg
         })
 
-class CustomerAPI(APIView): 
+class CustomerAPI1(APIView):
+
+    permission_classes = (IsAuthenticated, )
+
+    @swagger_auto_schema(responses=customer_get_response_all, 
+                         operation_description='Запросить список рассылок'
+                         )
+    def get(self, req):
+        return Response({
+            'customers': Customer.objects.all().values()
+            })
+    
+    @swagger_auto_schema(request_body=customer_post_request, 
+                         responses=customer_post_responses,
+                         operation_description='Добавление нового клиента для рассылки' 
+                         )
+    def post(self, req):
+        serializer = CustomerSerializer(data=req.data)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return json_response()
+
+
+class CustomerAPI2(APIView): 
 
     permission_classes = (IsAuthenticated, )
 
@@ -38,14 +65,29 @@ class CustomerAPI(APIView):
         try:
             return Customer.objects.get(pk=pk)
         except Customer.DoesNotExist:
-            return json_response('id does not exists', 0)
+            return None
 
-    def post(self, req):
-        serializer = CustomerSerializer(data=req.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return json_response()
+    @swagger_auto_schema(responses=customer_get_response, 
+                         operation_description='Запросить клиента по ID'
+                         )
     
+    def get(self, req, pk):
+        obj = self.get_object(pk)
+
+        if not obj:
+            return json_response('Клиент не найден!', 0)
+
+        return Response({
+            'id': obj.id,
+            'tag': obj.tag,
+            'tel':  obj.tel,
+            'tmz': obj.tmz,
+            })
+    
+    @swagger_auto_schema(responses=customer_put_responses, 
+                         operation_description='Внести изменения в рассылку',
+                         manual_parameters=[customer_par],
+                         )
     def put(self, req, pk):
         obj = self.get_object(pk)
         serializer = CustomerSerializer(instance=obj, data=req.data)
@@ -53,53 +95,129 @@ class CustomerAPI(APIView):
         serializer.save()
         return json_response()
     
+    @swagger_auto_schema(responses=customer_post_responses,
+                         operation_description='Удалить клиента по ID',
+                         manual_parameters=[customer_par],
+                         )
     def delete(self, req, pk):
-        print('ok')
         obj = self.get_object(pk)
+
+        if not obj:
+            return json_response('Клиент не найден!', 0)
+
         obj.delete()
         return json_response()
     
-class CampaignAPI(APIView): 
+
+class CampaignAPI1(APIView): 
 
     permission_classes = (IsAuthenticated, )
 
+    @swagger_auto_schema(responses=campaign_get_response_all, 
+                         operation_description='Запросить список рассылок'
+                         )
     def get(self, req):
         return Response({
             'campaigns': Campaign.objects.all().values()
             })
-    
-    def get_object(self, pk):
-        try:
-            return Campaign.objects.get(pk=pk)
-        except Campaign.DoesNotExist:
-            return json_response('нет рассылок', 0)
 
+    @swagger_auto_schema(request_body=campaign_post_request, 
+                         responses=campaign_post_responses,
+                         operation_description='Создание новой рассылки' 
+                         )
     def post(self, req):
         serializer = CampaignSerializer(data=req.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return json_response()
     
+
+class CampaignAPI2(APIView): 
+
+    permission_classes = (IsAuthenticated, )
+    
+    def get_object(self, pk):
+        try:
+            return Campaign.objects.get(pk=pk)
+        except Campaign.DoesNotExist:
+            return None
+
+    @swagger_auto_schema(responses=campaign_get_response, 
+                         operation_description='Запросить рассылку по ID'
+                         )
+    def get(self, req, pk):
+        obj = self.get_object(pk)
+
+        if not obj:
+            return json_response('Рассылка не найдена!', 0)
+
+        return Response({
+            'id': obj.id,
+            'message': obj.message,
+            'filter':  obj.filter,
+            'start': obj.start,
+            'stop': obj.stop,
+            'status': obj.status
+            })
+    
+    @swagger_auto_schema(request_body=campaign_put_request,
+                         responses=campaign_put_responses, 
+                         operation_description='Внести изменения в рассылку', 
+                         manual_parameters=[campaign_par]
+                         )
     def put(self, req, pk):
-
-        print(req.data)
-
         obj = self.get_object(pk)
         serializer = CampaignSerializer(instance=obj, data=req.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return json_response()
     
+    @swagger_auto_schema(responses=campaign_post_responses,
+                         operation_description='Удалить рассылку по ID',
+                         manual_parameters=[campaign_par],
+                         )
     def delete(self, req, pk):
         obj = self.get_object(pk)
+
+        if not obj:
+            return json_response('Рассылка не найдена!', 0)
+
         obj.delete()
-        print(obj)
         return json_response()
 
-class StatsData(APIView):
+
+class MessageAPI(APIView): 
+
+    permission_classes = (IsAuthenticated, )
+    
+    def get_object(self, pk):
+        try:
+            return MSG.objects.get(pk=pk)
+        except MSG.DoesNotExist:
+            return None
+
+    @swagger_auto_schema(responses=campaign_post_responses,
+                         operation_description='Удалить сообщение по ID',
+                         manual_parameters=[msg_par],
+                         )
+    
+    def delete(self, req, pk):
+        obj = self.get_object(pk)
+
+        if not obj:
+            return json_response('Сообщение не найдено!', 0)
+
+        obj.delete()
+        return json_response()
+    
+
+class StatsAPI1(APIView):
 
     permission_classes = (IsAuthenticated, )
 
+    @swagger_auto_schema(responses=stats_get_response_main,
+                         operation_description='Статистику по рассылкам',
+                         )
     def get(self, req):
 
         result = {}
@@ -114,19 +232,28 @@ class StatsData(APIView):
 
         return Response(result)
     
-class StatsDataByCampaign(APIView):
+
+class StatsAPI2(APIView):
 
     permission_classes = (IsAuthenticated, )
-
+    
     def get_object(self, pk):
         try:
             return Campaign.objects.get(pk=pk)
         except Campaign.DoesNotExist:
-            return json_response('нет рассылок', 0)
+            return None
 
+    @swagger_auto_schema(responses=stats_get_response,
+                         operation_description='Запросить статистику по ID рассылки',
+                         manual_parameters=[campaign_par],
+                         )
+    
     def get(self, req, pk):
 
         camp = self.get_object(pk)
+
+        if not camp:
+            return json_response('Рассылка не найдена!', 0)
 
         status = ''
 
@@ -136,17 +263,19 @@ class StatsDataByCampaign(APIView):
             status = 'запущенные'
         elif camp.status == 3:
             status = 'завершенные'
-
-        msgs = MSG.objects.filter(campaign_id=pk)
-
+                
         succeed = 0
         failed = 0
 
-        for msg in msgs:
-            if msg.status == 200:
-                succeed += 1
-            else:
-                failed += 1
+        msgs = MSG.objects.filter(campaign_id=pk)
+
+        if msgs:
+            
+            for msg in msgs:
+                if msg.status == 200:
+                    succeed += 1
+                else:
+                    failed += 1
 
         return Response({
             'campaign_id': pk,
@@ -156,3 +285,4 @@ class StatsDataByCampaign(APIView):
                 'failed': failed
             }
             })
+
